Use hash instead of a calculated id for file name.
Previous way of ID calculation is a result of German course
analysis, where audio files have the following style of links:
.../id/1.mp3
As it appeared in Swedish it's different (in other languages
probably too). That's why it was decided to use MD5 hash
(calculated from audios URL) to give a unique ID to each file.